修改了几个文件

+++ Controller+/GuidePage/GuidepageCollectionCell.h
@@ -2,15 +2,15 @@
 //  GuidepageCollectionCell.h
 //  OCCommon
 //
-//  Created by 一泓明峰 on 16/3/5.
-//  Copyright © 2016年 RJ. All rights reserved.
+//  Created by 游峰 on 16/3/5.
+//  Copyright © 2016年 yf. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>
 
 @interface GuidepageCollectionCell : UICollectionViewCell
 @property (weak, nonatomic) UIImage * image;
-- (void)setIndexPath:(NSIndexPath *)indexPath count:(int)count;
+- (void)setIndexPath:(NSIndexPath *)indexPath count:(NSInteger)count;
 + (instancetype)guidepageCollectionCellWithCcollectionView:(UICollectionView *)collectionView cellForItemAtIndexPath:(NSIndexPath *)indexPath;
 
 @end
