+++ Controller+/GuidePage/GuidepageViewController.h
@@ -2,14 +2,14 @@
 //  GuidepageViewController.h
 //  OCCommon
 //
-//  Created by 一泓明峰 on 16/3/4.
-//  Copyright © 2016年 RJ. All rights reserved.
+//  Created by 游峰 on 16/3/4.
+//  Copyright © 2016年 yf. All rights reserved.
 //
 
 #import "BaseViewController.h"
 
 @interface GuidepageViewController : BaseViewController
 
-@property (assign, nonatomic) int kGuidepageImageCount;
+@property (assign, nonatomic) NSInteger kGuidepageImageCount;
 
 @end
